Squashed commit of the following:
commit f16a294b2f8a6802933d21d048010270b617ab79
Author: makian <hohosmarko@gmail.com>
Date:   Sat Aug 3 19:08:21 2024 +0200

    Gravity changed on world

commit a244f89349ce664633bdf7493120c198bca572a2
Author: makian <hohosmarko@gmail.com>
Date:   Sat Aug 3 19:02:38 2024 +0200

    Added world and input manager updated, fixed ship.hpp not including block

+++ src/inputmanager.cpp
@@ -8,17 +8,13 @@ void InputManager::Update() {
 }
 void InputManager::OnKeyPress(sf::Keyboard::Scancode key)
 {
-	keysDown[key] = true;
+	keysDown[static_cast<std::uint32_t>(key)] = true;
 }
 void InputManager::OnKeyRelease(sf::Keyboard::Scancode key)
-{
-	keysDown[key] = false;
-}
+{ keysDown[static_cast<std::uint32_t>(key)] = false; }
 bool InputManager::IsKeyDown(sf::Keyboard::Scancode key) const
-{
-	return keysDown.at(key);
-}
+{ return keysDown.at(static_cast<std::uint32_t>(key)); }
 bool InputManager::IsKeyPressed(sf::Keyboard::Scancode key) const
 {
-	return keysDown.at(key) && !keysLastDown.at(key);
+	return keysDown.at(static_cast<std::uint32_t>(key)) && !keysLastDown.at(static_cast<std::uint32_t>(key));
 }
